Implement the add product feature

+++ Product.cs
@@ -3,14 +3,14 @@
 public class Product
 {
     private string name = string.Empty;
-    private float price;
+    private double price;
     private int quantity;
     
     public string Name { get => name; set => name = value; }
-    public float Price { get => price; set => price = value; }
+    public double Price { get => price; set => price = value; }
     public int Quantity { get => quantity; set => quantity = value; }
 
-    public Product(string name, float price, int quantity)
+    public Product(string name, double price, int quantity)
     {
         this.Name = name;
         this.Price = price;
